+++ web/apps/photos/src/components/PhotoViewer/FileInfo/index.tsx
@@ -10,11 +10,8 @@ import TextSnippetOutlined from "@mui/icons-material/TextSnippetOutlined";
 import { Box, DialogProps, Link, Stack, styled } from "@mui/material";
 import { Chip } from "components/Chip";
 import { EnteDrawer } from "components/EnteDrawer";
-import {
-    PhotoPeopleList,
-    UnidentifiedFaces,
-} from "components/MachineLearning/PeopleList";
 import Titlebar from "components/Titlebar";
+import { PhotoPeopleList, UnidentifiedFaces } from "components/ml/PeopleList";
 import LinkButton from "components/pages/gallery/LinkButton";
 import { t } from "i18next";
 import { AppContext } from "pages/_app";
