+++ web/apps/photos/src/components/Search/SearchBar/searchInput/MenuWithPeople.tsx
@@ -1,6 +1,6 @@
 import { Row } from "@ente/shared/components/Container";
 import { Box, styled } from "@mui/material";
-import { PeopleList } from "components/MachineLearning/PeopleList";
+import { PeopleList } from "components/ml/PeopleList";
 import { t } from "i18next";
 import { AppContext } from "pages/_app";
 import { useContext } from "react";
